Move signing classes to a signer package.
