+++ validator/client/src/main/java/tech/pegasys/artemis/validator/client/signer/ExternalSignerException.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.artemis.validator.client;
+package tech.pegasys.artemis.validator.client.signer;
 
 public class ExternalSignerException extends RuntimeException {
