+++ validator/coordinator/src/main/java/tech/pegasys/artemis/validator/coordinator/ValidatorInfo.java
@@ -13,7 +13,7 @@
 
 package tech.pegasys.artemis.validator.coordinator;
 
-import tech.pegasys.artemis.validator.client.Signer;
+import tech.pegasys.artemis.validator.client.signer.Signer;
 
 public class ValidatorInfo {
